# PERF: Speed up Spearman calculation