--- a/asv_bench/benchmarks/stat_ops.py
+++ b/asv_bench/benchmarks/stat_ops.py
@@ -113,8 +113,8 @@ class Correlation:
             nanops._USE_BOTTLENECK = use_bottleneck
         self.df = pd.DataFrame(np.random.randn(1000, 30))
         self.df2 = pd.DataFrame(np.random.randn(1000, 30))
-        self.df_wide = pd.DataFrame(np.random.randn(1000, 500))
-        self.df_wide_nans = self.df_wide.where(np.random.random((1000, 500)) < 0.9)
+        self.df_wide = pd.DataFrame(np.random.randn(1000, 300))
+        self.df_wide_nans = self.df_wide.where(np.random.random((1000, 300)) < 0.9)
         self.s = pd.Series(np.random.randn(1000))
         self.s2 = pd.Series(np.random.randn(1000))
 
@@ -122,15 +122,9 @@ class Correlation:
         self.df.corr(method=method)
 
     def time_corr_wide(self, method, use_bottleneck):
-        # Raise here until benchmark can pass
-        if method == "kendall":
-            return None
         self.df_wide.corr(method=method)
 
     def time_corr_wide_nans(self, method, use_bottleneck):
-        # Raise here until benchmark can pass
-        if method == "kendall":
-            return None
         self.df_wide_nans.corr(method=method)
 
     def peakmem_corr_wide(self, method, use_bottleneck):