--- a/pandas/_libs/algos.pyx
+++ b/pandas/_libs/algos.pyx
@@ -316,13 +316,13 @@ def nancorr_spearman(const float64_t[:, :] mat, Py_ssize_t minp=1):
             nobs = 0
             # Keep track of whether the two columns have the same
             # missing pattern, if so save off the ranks
-            same_miss_pat = True
+            all_ranks = True
             for i in range(N):
-                same_miss_pat &= not (mask[i, xi] ^ mask[i, yi])
+                all_ranks &= not (mask[i, xi] ^ mask[i, yi])
                 if mask[i, xi] and mask[i, yi]:
                     nobs += 1
 
-            if same_miss_pat:
+            if all_ranks:
                 if xi not in cached_ranks:
                     ranked_mat[:, xi] = rank_1d_float64(mat[:, xi])
                     cached_ranks.add(xi)
@@ -339,11 +339,11 @@ def nancorr_spearman(const float64_t[:, :] mat, Py_ssize_t minp=1):
 
                 for i in range(N):
                     if mask[i, xi] and mask[i, yi]:
-                        maskedx[j] = ranked_mat[i, xi] if same_miss_pat else mat[i, xi]
-                        maskedy[j] = ranked_mat[i, yi] if same_miss_pat else mat[i, yi]
+                        maskedx[j] = ranked_mat[i, xi] if all_ranks else mat[i, xi]
+                        maskedy[j] = ranked_mat[i, yi] if all_ranks else mat[i, yi]
                         j += 1
 
-                if not same_miss_pat:
+                if not all_ranks:
                     maskedx = rank_1d_float64(maskedx)
                     maskedy = rank_1d_float64(maskedy)