@f-vert 2019-08-27
does data analyst need a bachelors ?

On Mon, Aug 26, 2019 at 3:32 PM Daniel Saxton <notifications@github.com>
wrote:

> so in some situations this will actually perform slightly worse than the
> current implementation
>
> Is there a good heuristic that we can use to check this?
>
> The worst case would be when every column has a different missing pattern
> and the first step of precalculating is for naught, and in this case you do
> an extra k sorts. I think this is a good tradeoff since in the best case
> you do k instead of k**2 sorts.
>
> When I run that scenario on this branch I do get worse performance (4.74
> seconds vs. 4.49 seconds using a 5000 by 100 array), but it's actually
> almost within a margin of error of the current implementation.
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/pandas-dev/pandas/pull/28151?email_source=notifications&email_token=AKHFDSCIQMRPYRIYLB6NOLLQGRKYLA5CNFSM4IPQM5WKYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOD5F4FCA#issuecomment-525058696>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AKHFDSB2X4JPHIYCT4KU3VDQGRKYLANCNFSM4IPQM5WA>
> .
>