@WillAyd 2019-09-05
You might need to skip the Kendall method in benchmarks where failing to get CI to pass. can you open up a follow up issue to address?