@WillAyd 2019-09-05
Try raising a `NotImplementedError`