@dsaxton 2019-09-06
> Try raising a `NotImplementedError`

I was seeing the benchmark still fail after doing this.  Feels like a hack, but how about returning `None` (or maybe making the data just small enough so they all pass)?